pyproject.toml
[project]
name = "mpt-ai-e2e-tester"
version = "0.1.0"
description = "MPT AI E2E Tester - Test the MPT AI platform with automated scenarios"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=11.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mpt-cli = "main:main"
mpt-mock-server = "mock_voice_manager:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["B011"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]


